fixes on projections and queries api
