+++ libraries/python/com-ivansoft-corebank-lib/com_ivansoft_corebank_lib/models/Balance.py
@@ -4,7 +4,7 @@ from decimal import Decimal
 from typing import Optional
 
 
-class TransactionModel(BaseModel):
+class Balance(BaseModel):
     balance: Decimal
     currency: str = 'MXN'
     user_id: int
