+++ libraries/python/com-ivansoft-corebank-lib/com_ivansoft_corebank_lib/models/User.py
@@ -1,6 +1,6 @@
 from pydantic import BaseModel
 
 
-class UserModel(BaseModel):
+class User(BaseModel):
     user_id: int
     username: str
