+++ libraries/python/com-ivansoft-corebank-lib/pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "com-ivansoft-corebank-lib"
-version = "0.1.49"
+version = "0.1.54"
 description = ""
 authors = ["Rodrigo Zamora"]
 readme = "README.md"
