+++ projections/account/app/db/balance/BalanceRepository.py
@@ -1,5 +1,5 @@
 import json
-from app.db.balance.BalanceModel import BalanceModel
+from com_ivansoft_corebank_lib.models.Balance import Balance as BalanceModel
 from structlog import get_logger
 from motor.motor_asyncio import AsyncIOMotorClient
 from app.db.MongoBase import MongoBase
@@ -24,12 +24,3 @@ class BalanceRepository:
                          .find_one({'account_id': account_id}, sort=[('updated_at', -1)]))
 
         return BalanceModel(**balance) if balance else None
-
-    async def get_history(self, account_id: str):
-        logger.info('Retrieving balance history', account_id=account_id)
-
-        # find all balances, order by updated_at field
-        history = await (BalanceRepository._client[MONGO_DB_NAME][MONGO_BALANCE_COLLECTION]
-                         .find({'account_id': account_id}).sort('updated_at', -1).to_list(length=None))
-
-        return [BalanceModel(**balance) for balance in history]
