--- projections/account/app/db/transaction/TransactionModel.py
@@ -1,20 +1,0 @@
-
-from pydantic import BaseModel
-from decimal import Decimal
-from typing import Optional
-
-
-class TransactionModel(BaseModel):
-    balance: Decimal
-    currency: str = 'MXN'
-    user_id: int
-    username: str
-    account_id: str
-    # optional fields
-    created_at: Optional[str] = None
-    updated_at: Optional[str] = None
-
-    class Config:
-        json_encoders = {
-            Decimal: lambda v: str(v)
-        }
