+++ projections/account/app/db/transaction/TransactionRepository.py
@@ -1,5 +1,5 @@
 import json
-from app.db.transaction.TransactionModel import TransactionModel
+from com_ivansoft_corebank_lib.models.Transaction import Transaction as TransactionModel
 from structlog import get_logger
 from motor.motor_asyncio import AsyncIOMotorClient
 from app.db.MongoBase import MongoBase
@@ -15,21 +15,3 @@ class TransactionRepository:
         to_save = json.loads(transaction.model_dump_json())
         logger.info('Saving transaction', balance=transaction)
         await TransactionRepository._client[MONGO_DB_NAME][MONGO_TRANSACTION_COLLECTION].insert_one(to_save)
-
-    async def get(self, account_id: str) -> TransactionModel:
-        logger.info('Retrieving transaction', account_id=account_id)
-
-        # Get last balance, order by updated_at field
-        transaction = await (TransactionRepository._client[MONGO_DB_NAME][MONGO_TRANSACTION_COLLECTION]
-                         .find_one({'account_id': account_id}, sort=[('updated_at', -1)]))
-
-        return TransactionModel(**transaction) if transaction else None
-
-    async def get_history(self, account_id: str):
-        logger.info('Retrieving transaction history', account_id=account_id)
-
-        # find all balances, order by updated_at field
-        history = await (TransactionRepository._client[MONGO_DB_NAME][MONGO_TRANSACTION_COLLECTION]
-                         .find({'account_id': account_id}).sort('updated_at', -1).to_list(length=None))
-
-        return [TransactionModel(**transaction) for transaction in history]
