--- projections/account/app/db/user/UserModel.py
@@ -1,6 +1,0 @@
-from pydantic import BaseModel
-
-
-class UserModel(BaseModel):
-    user_id: int
-    username: str
