+++ queries_bank_api/app/db/balance/BalanceRepository.py
@@ -1,5 +1,5 @@
-import json
-from app.db.balance.BalanceModel import BalanceModel
+
+from com_ivansoft_corebank_lib.models.Balance import Balance as BalanceModel
 from structlog import get_logger
 from motor.motor_asyncio import AsyncIOMotorClient
 from app.db.MongoBase import MongoBase
@@ -10,11 +10,6 @@ logger = get_logger().bind(logger='BalanceRepository')
 
 class BalanceRepository:
     _client: AsyncIOMotorClient = MongoBase.get_client()
-
-    async def save(self, balance: BalanceModel):
-        to_save = json.loads(balance.model_dump_json())
-        logger.info('Saving balance', balance=balance)
-        await BalanceRepository._client[MONGO_DB_NAME][MONGO_BALANCE_COLLECTION].insert_one(to_save)
 
     async def get(self, account_id: str) -> BalanceModel:
         logger.info('Retrieving balance', account_id=account_id)
