+++ queries_bank_api/app/db/transaction/TransactionRepository.py
@@ -1,5 +1,5 @@
 import json
-from app.db.transaction.TransactionModel import TransactionModel
+from com_ivansoft_corebank_lib.models.Transaction import Transaction as TransactionModel
 from structlog import get_logger
 from motor.motor_asyncio import AsyncIOMotorClient
 from app.db.MongoBase import MongoBase
@@ -10,11 +10,6 @@ logger = get_logger().bind(logger='TransactionRepository')
 
 class TransactionRepository:
     _client: AsyncIOMotorClient = MongoBase.get_client()
-
-    async def save(self, transaction: TransactionModel):
-        to_save = json.loads(transaction.model_dump_json())
-        logger.info('Saving transaction', balance=transaction)
-        await TransactionRepository._client[MONGO_DB_NAME][MONGO_TRANSACTION_COLLECTION].insert_one(to_save)
 
     async def get(self, account_id: str) -> TransactionModel:
         logger.info('Retrieving transaction', account_id=account_id)
